refactor: 목록에 사용되는 컴포넌트의 디렉토리 위치 수정
