+++ client/src/pages/Category/Exhibition/Exhibitions/Exhibitions.jsx
@@ -2,10 +2,10 @@ import styles from './Exhibitions.module.css';
 import { useEffect, useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { instance } from '../../../../apis/instance.js';
-import ListHeader from '../../../../components/List/ListHeader/ListHeader';
-import DropdownContainer from '../../../../components/List/DropdownContainer/DropdownContainer';
+import ListHeader from '../../components/ListHeader/ListHeader';
+import DropdownContainer from '../../components/DropdownContainer/DropdownContainer';
 import { exhibitionFilter } from '../../../../utils/filters/exhibitionFilter.js';
-import TotalCounts from '../../../../components/List/TotalCounts/TotalCounts';
+import TotalCounts from '../../components/TotalCounts/TotalCounts';
 import Pagination from '../../../../components/Pagination/Pagination';
 import usePagination from '../../../../hooks/usePagination';
 import ExhibitionCard from './components/ExhibitionCard/ExhibitionCard';
