+++ client/src/pages/Category/components/TotalCounts/TotalCounts.jsx
